requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-js-routes==0.2.0
django-vite==3.0.3
idna==3.6
inertia-django==0.6.0
phonenumbers==8.13.30
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
